third_party/blink/web_tests/svg/dynamic-updates/SVGFEPointLightElement-dom-x-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' attribute of the SVGFEPointLightElement object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.getAttribute('x') is "50"
PASS pointLight.getAttribute('x') is "100"
PASS successfullyParsed is true

TEST COMPLETE
